refactor(shared-session): shared_session_v2 主体拆为五域子模块加聚合层（5349→870）
